astro/extractedcode/example-client-libraries/python/requirements.txt
fusionauth-client==1.42.0
